refactor: ApiError.java to new and clearer name, remove old exception that is now handled by the global handler
