+++ src/main/java/com/lisaanna/ws_todo/exception/GlobalExceptionHandler.java
@@ -1,6 +1,5 @@
 package com.lisaanna.ws_todo.exception;
 
-import com.lisaanna.ws_todo.controller.ApiError;
 import io.github.resilience4j.ratelimiter.RequestNotPermitted;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
@@ -8,19 +7,16 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
-import java.util.HashMap;
-import java.util.Map;
-
 @ControllerAdvice
 public class GlobalExceptionHandler {
 
-    private ResponseEntity<ApiError> buildResponse(
+    private ResponseEntity<ErrorResponseBody> buildResponse(
             HttpStatus status,
             String error,
             String message,
             HttpServletRequest request
     ) {
-        ApiError apiError = new ApiError(
+        ErrorResponseBody apiError = new ErrorResponseBody(
                 status.value(),
                 error,
                 message,
@@ -30,7 +26,7 @@ public class GlobalExceptionHandler {
     }
 
     @ExceptionHandler(RequestNotPermitted.class)
-    public ResponseEntity<ApiError> handleRequestNotPermitted(RequestNotPermitted e, HttpServletRequest request) {
+    public ResponseEntity<ErrorResponseBody> handleRequestNotPermitted(RequestNotPermitted e, HttpServletRequest request) {
         return buildResponse(
                 HttpStatus.TOO_MANY_REQUESTS,
                 "Too Many Requests",
@@ -40,7 +36,7 @@ public class GlobalExceptionHandler {
     }
 
     @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<ApiError> handleRuntimeException(RuntimeException e, HttpServletRequest request) {
+    public ResponseEntity<ErrorResponseBody> handleRuntimeException(RuntimeException e, HttpServletRequest request) {
 
         return buildResponse(
                 HttpStatus.INTERNAL_SERVER_ERROR,
@@ -51,7 +47,7 @@ public class GlobalExceptionHandler {
     }
 
     @ExceptionHandler(NullPointerException.class)
-    public ResponseEntity<ApiError> handleNullPointerException(NullPointerException e, HttpServletRequest request) {
+    public ResponseEntity<ErrorResponseBody> handleNullPointerException(NullPointerException e, HttpServletRequest request) {
         return buildResponse(
                 HttpStatus.INTERNAL_SERVER_ERROR,
                 "Null Pointer Exception",
